Inspect command supports read volume/backup metadata
- backup inspect-volume <volume>: read a backup volume metadata
- backup inspect <backup>: read a backup metadata
